Replaced the compiler macro with an ordinary one.
Always require the second argument to COMPILE-XPATH.  We cannot see lexical
bindings in this function.

Replaced the function EVALUATE with EVALUATE-THUNK, which always takes a
closure as an argument, never a string or sexp.  (EVALUATE-THUNK is just
a working title for this function, perhaps a better name can be found.)

New macro EVALUATE, replacing the compiler macro.

Removed *LEXICAL-NAMESPACES* and *LEXICAL-VARIABLES*.
  - don't bind *LEXICAL-NAMESPACES* in WITH-NAMESPACES
  - don't bind *LEXICAL-VARIABLES* in WITH-VARIABLES

Export EVALUATE-THUNK from the package in addition to EVALUATE.

Added a test checking that the empty namespace can be used without
having to use WITH-NAMESPACES.

Removed the call to STRING in TEST-FILTERING.  (fixme?)

darcs-hash:20071220214420-e47ce-55b42a4c50757f63616b05dccca0cac08aa84fb4.gz